+++ src/app/locales/en/components/file-metadata/-common.js
@@ -1,6 +1,6 @@
 export const noMetadataSet = 'There is no {{type}} metadata set.';
 export const types = {
-  xattrs: 'basic',
+  xattrs: 'xattrs',
   json: 'JSON',
   rdf: 'RDF',
 };