+++ src/app/locales/en/components/file-metadata/header.js
@@ -12,4 +12,11 @@ export default {
     validating: '{{typeCapitalized}} metadata validation in progress...',
   },
   editorDisabled: 'Editor disabled',
+  selectorHint: {
+    viewDescription: 'This view allows managing custom metadata associated with files and directories.',
+    metadataIntro: 'Every file or directory can have three types of metadata: extended attributes, JSON, and RDF. These can be set automatically (e.g., by scripts, middleware, or automation jobs) or manually by users. Custom metadata enables rich annotation of data and has various applications, such as provenance tracking, semantic search, data governance, or geospatial context mapping.',
+    metadataAdditionalInfo: 'Notably, custom metadata facilitates data indexing and discovery, which is especially useful for managing large data collections.',
+    docLinkName: 'metadata',
+    close: 'OK',
+  },
 };